% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medbayes.R, R/medbayes_ZI_mediator.R,
%   R/medbayes_ZI_outcome.R
\name{medbayes}
\alias{medbayes}
\alias{medbayes_zim}
\alias{medbayes_ziy}
\title{Run Bayesian mediation model}
\usage{
medbayes(
  model.m = model.m,
  model.y = model.y,
  treat = "treatment",
  mediator = "mediator",
  ind_mediator = NULL,
  outcome = "outcome",
  control.value = 0,
  treat.value = 1
)

medbayes_zim(
  model.m = model.m,
  model.y = model.y,
  treat = "treatment",
  mediator = "mediator",
  ind_mediator = NULL,
  outcome = "outcome",
  control.value = 0,
  treat.value = 1,
  logM = FALSE
)

medbayes_ziy(
  model.m = model.m,
  model.y = model.y,
  treat = "treatment",
  mediator = "mediator",
  ind_mediator = NULL,
  outcome = "outcome",
  control.value = 0,
  treat.value = 1
)
}
\arguments{
\item{model.m}{A brms model for the mediator}

\item{model.y}{A brms model for the outcome}

\item{dat.new}{New data for prediction}
}
\value{
Posterior predictions (first column)

Posterior predictions (first column)

Posterior predictions (first column)
}
\description{
Run Bayesian mediation model

Run Bayesian mediation model

Run Bayesian mediation model
}
